Updating ulps_eq into a cleaner module
